+++ src/logging.ts
@@ -5,8 +5,14 @@ export type { Logger } from 'pino';
 export function createDefaultLogger() {
   return Pino({
     serializers: Pino.stdSerializers,
-    name: '@nostalgie/server',
-    prettyPrint: process.env.NODE_ENV !== 'production',
+    name: '@nostalgie/internal/server',
+    prettyPrint:
+      process.env.NODE_ENV !== 'production'
+        ? {
+            ignore: 'name,hostname,pid',
+            translateTime: 'HH:MM:ss.l',
+          }
+        : false,
     redact: ['req.headers'],
     timestamp: Pino.stdTimeFunctions.isoTime,
   });
