+++ src/runtime/internal/server/renderer.tsx
@@ -10,15 +10,15 @@ import { install } from 'source-map-support';
 // import StackTracey from 'stacktracey';
 import * as Twind from 'twind';
 import * as TwindServer from 'twind/server';
-import type { ChunkDependencies } from '../../build/types';
-import type { ClientAuth, NostalgieUser } from '../auth';
-import { AuthContext, ServerAuth } from '../auth/server';
+import type { ChunkDependencies } from '../../../build/types';
+import type { ClientAuth, NostalgieUser } from '../../auth';
+import { AuthContext, ServerAuth } from '../../auth/server';
+import { ServerQueryContextProvider, ServerQueryExecutorImpl } from '../../functions/server';
+import type { ServerFunction, ServerFunctionContext } from '../../functions/types';
+import { defaultHelmetProps } from '../../helmet';
+import { LazyContext } from '../../lazy/context';
+import type { ChunkManager } from '../../lazy/types';
 import type { BootstrapOptions } from '../bootstrap/bootstrap';
-import { ServerQueryContextProvider, ServerQueryExecutorImpl } from '../functions/server';
-import type { ServerFunction, ServerFunctionContext } from '../functions/types';
-import { defaultHelmetProps } from '../helmet';
-import { LazyContext } from '../lazy/context';
-import type { ChunkManager } from '../lazy/types';
 
 if (process.env.NODE_ENV === 'development') {
   install({
